Create daily_commit_20210421_python_smart_dictionary.py
Python Smart Dictionary:
Given a words.txt file containing an English-Chinese words list, design a fast and smart dictionary tool for Chinese learner. This will involve first reading all items from the words.txt which is a text file containing a list of English words and their corresponding Chinese translation.

Basic and extended functions:
a) The user should be able to input letters and the dictionary return the match words list and its match count. b) The user can select a word to show the its Chinese counterpart. This includes all important information like the word type as well as the Chinese characters
c) The user can enter Chinese word and show all possible matched English words list.
d) Count the number of occurrences of 26 letters in the dictionary, and draw the occurrences with bar graph as
e) Design a GUI for it (optional)